.history/requirements_20251106210224.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
drf-yasg==1.21.4
Pillow==9.5.0
django-filter==22.1
gunicorn==20.1.0
whitenoise==6.4.0